requirements.txt
vk-api==11.9.9
Flask==3.0.0
python-dotenv==1.0.0